+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
+* Simple API to specify custom rules using XPath for XML comparison
 * Change: Custom json matchers are only tested against the new actual result and not against the persisted snapshot anymore
 * [#24](https://github.com/skuzzle/snapshot-tests/issues/24): Only create snapshot directory if necessary